# nepflow/src/main/java/com/nepflow/NepenthesManagement/Controller/SpeciesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.0.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.nepflow.NepenthesManagement.Controller;

import com.nepflow.NepenthesManagement.Dto.LabelClonesDTO;
import com.nepflow.NepenthesManagement.Dto.LabelDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-06-11T01:56:15.806697200+02:00[Europe/Berlin]")
@Validated
@Tag(name = "Species", description = "the Species API")
public interface SpeciesApi {

    default SpeciesApiDelegate getDelegate() {
        return new SpeciesApiDelegate() {};
    }

    /**
     * GET /clone/species/ : get all Species
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneSpeciesGet",
        summary = "get all Species",
        tags = { "Species" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = LabelDTO.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/clone/species/",
        produces = { "application/json" }
    )
    default ResponseEntity<List<LabelDTO>> cloneSpeciesGet(
        
    ) {
        return getDelegate().cloneSpeciesGet();
    }


    /**
     * GET /clone/species/ic : get all IC Species Clones
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneSpeciesIcGet",
        summary = "get all IC Species Clones",
        tags = { "Species" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = LabelClonesDTO.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/clone/species/ic",
        produces = { "application/json" }
    )
    default ResponseEntity<List<LabelClonesDTO>> cloneSpeciesIcGet(
        
    ) {
        return getDelegate().cloneSpeciesIcGet();
    }


    /**
     * GET /clone/species/ic/{name} : get all IC Species Clones of a Nepenthes
     *
     * @param name  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneSpeciesIcNameGet",
        summary = "get all IC Species Clones of a Nepenthes",
        tags = { "Species" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/clone/species/ic/{name}",
        produces = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneSpeciesIcNameGet(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name
    ) {
        return getDelegate().cloneSpeciesIcNameGet(name);
    }


    /**
     * POST /clone/species/ic/{name} : add a new iv clone to a nepenthes
     *
     * @param name  (required)
     * @param labelClonesDTO Clone DTO with new values (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneSpeciesIcNamePost",
        summary = "add a new iv clone to a nepenthes",
        tags = { "Species" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/clone/species/ic/{name}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneSpeciesIcNamePost(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name,
        @Parameter(name = "LabelClonesDTO", description = "Clone DTO with new values", required = true) @Valid @RequestBody LabelClonesDTO labelClonesDTO
    ) {
        return getDelegate().cloneSpeciesIcNamePost(name, labelClonesDTO);
    }


    /**
     * PUT /clone/species/ic/{name}
     *
     * @param name  (required)
     * @param labelClonesDTO  (required)
     * @return OK (status code 200)
     *         or No nepenthes found (status code 404)
     */
    @Operation(
        operationId = "cloneSpeciesIcNamePut",
        tags = { "Species" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "No nepenthes found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/clone/species/ic/{name}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneSpeciesIcNamePut(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name,
        @Parameter(name = "LabelClonesDTO", description = "", required = true) @Valid @RequestBody LabelClonesDTO labelClonesDTO
    ) {
        return getDelegate().cloneSpeciesIcNamePut(name, labelClonesDTO);
    }


    /**
     * GET /clone/species/iv/{name} : get all IC Species Clones of a Nepenthes
     *
     * @param name  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneSpeciesIvNameGet",
        summary = "get all IC Species Clones of a Nepenthes",
        tags = { "Species" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/clone/species/iv/{name}",
        produces = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneSpeciesIvNameGet(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name
    ) {
        return getDelegate().cloneSpeciesIvNameGet(name);
    }


    /**
     * POST /clone/species/iv/{name} : add a new iv clone to a nepenthes
     *
     * @param name  (required)
     * @param labelClonesDTO Clone DTO with new values (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneSpeciesIvNamePost",
        summary = "add a new iv clone to a nepenthes",
        tags = { "Species" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/clone/species/iv/{name}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneSpeciesIvNamePost(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name,
        @Parameter(name = "LabelClonesDTO", description = "Clone DTO with new values", required = true) @Valid @RequestBody LabelClonesDTO labelClonesDTO
    ) {
        return getDelegate().cloneSpeciesIvNamePost(name, labelClonesDTO);
    }


    /**
     * PUT /clone/species/iv/{name}
     *
     * @param name  (required)
     * @param labelClonesDTO  (required)
     * @return OK (status code 200)
     *         or No nepenthes found (status code 404)
     */
    @Operation(
        operationId = "cloneSpeciesIvNamePut",
        tags = { "Species" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "No nepenthes found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/clone/species/iv/{name}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneSpeciesIvNamePut(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name,
        @Parameter(name = "LabelClonesDTO", description = "", required = true) @Valid @RequestBody LabelClonesDTO labelClonesDTO
    ) {
        return getDelegate().cloneSpeciesIvNamePut(name, labelClonesDTO);
    }


    /**
     * GET /clone/species/{name} : get a Species and their clones
     *
     * @param name  (required)
     * @return OK (status code 200)
     *         or nepenthes not found (status code 404)
     */
    @Operation(
        operationId = "cloneSpeciesNameGet",
        summary = "get a Species and their clones",
        tags = { "Species" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "nepenthes not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/clone/species/{name}",
        produces = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneSpeciesNameGet(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name
    ) {
        return getDelegate().cloneSpeciesNameGet(name);
    }


    /**
     * POST /clone/species/ : add a new Species
     *
     * @param labelClonesDTO  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneSpeciesPost",
        summary = "add a new Species",
        tags = { "Species" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/clone/species/",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneSpeciesPost(
        @Parameter(name = "LabelClonesDTO", description = "", required = true) @Valid @RequestBody LabelClonesDTO labelClonesDTO
    ) {
        return getDelegate().cloneSpeciesPost(labelClonesDTO);
    }

}