# nepflow/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: API for Nepenthes Operations
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Nepflow Nepenthes
  version: 1.0.0
servers:
- url: http://localhost:8080/api
...
paths:
...
  /clone/species/:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  allOf:
                  - $ref: '#/components/schemas/LabelDTO'
                type: array
          description: OK
      summary: get all Species
      tags:
      - Species
      x-accepts: application/json
      x-tags:
      - tag: Species
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelClonesDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
      summary: add a new Species
      tags:
      - Species
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Species
  /clone/species/{name}:
    get:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
        "404":
          description: nepenthes not found
      summary: get a Species and their clones
      tags:
      - Species
      x-accepts: application/json
      x-tags:
      - tag: Species
  /clone/species/iv/{name}:
    get:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
      summary: get all IC Species Clones of a Nepenthes
      tags:
      - Species
      x-accepts: application/json
      x-tags:
      - tag: Species
    post:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelClonesDTO'
        description: Clone DTO with new values
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
      summary: add a new iv clone to a nepenthes
      tags:
      - Species
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Species
    put:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelClonesDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
        "404":
          description: No nepenthes found
      tags:
      - Species
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Species
...
  /clone/species/ic/{name}:
    get:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
      summary: get all IC Species Clones of a Nepenthes
      tags:
      - Species
      x-accepts: application/json
      x-tags:
      - tag: Species
    post:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelClonesDTO'
        description: Clone DTO with new values
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
      summary: add a new iv clone to a nepenthes
      tags:
      - Species
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Species
    put:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelClonesDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
        "404":
          description: No nepenthes found
      tags:
      - Species
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Species
...
components:
  schemas:
...
    LabelDTO:
      example:
        prefix: prefix
        name: name
      properties:
        name:
          type: string
        prefix:
          type: string
      type: object
    CloneDTO:
      properties:
        cloneId:
          type: string
        internalcloneId:
          type: string
        sex:
          type: string
        grex:
          $ref: '#/components/schemas/CloneGrexDTO'
        producer:
          type: string
        Location:
          type: string
      type: object
    HybridCloneDTO:
      allOf:
      - $ref: '#/components/schemas/CloneDTO'
      - properties:
          fatherName:
            type: string
          motherName:
            type: string
        type: object
    CloneGrexDTO:
      properties:
        cloneId:
          type: string
        father:
          $ref: '#/components/schemas/CloneGrexDTO_father'
        mother:
          $ref: '#/components/schemas/CloneGrexDTO_father'
      type: object
    LabelClonesDTO:
      example:
        clones:
        - null
        - null
        label:
          prefix: prefix
          name: name
      properties:
        label:
          $ref: '#/components/schemas/LabelDTO'
        clones:
          items:
            $ref: '#/components/schemas/LabelClonesDTO_clones_inner'
          type: array
      type: object
    CloneGrexDTO_father:
      oneOf:
      - $ref: '#/components/schemas/CloneDTO'
      - $ref: '#/components/schemas/HybridCloneDTO'
      x-one-of-name: CloneGrexDTOFather
    LabelClonesDTO_clones_inner:
      anyOf:
      - $ref: '#/components/schemas/CloneDTO'
      - $ref: '#/components/schemas/HybridCloneDTO'